* master: (115 commits)
  Mack 0.5.0 release
  Fixed bad environment lookup.
  Added erubis 2.6.0
  forgot to submit this file with my previous submission...
  Updated CHANGELOG
  Commented out cachetastic:page_cache:expire_all task since, currently there is no page caching.
  rake console now just loads the console.rb file.
  Refactored out constants, such as MACK_ROOT and MACK_ENV and made them into Mack::Configuration.env, Mack::Configuration.root, etc...
  oops, forgot to update CHANGELOG
  added a test case for rendering 404 page
  render() now accepts :status option.
  Merge commit 'origin'; commit 'coreteam/master'
  set mack-data_mapper gem version equal to the current gem version.
  Removed dependency on the data_mapper gem to the mack-data_mapper gem.
  Fixed a bug loading url.rb and uploaded_file.rb found by Phil Darnowsky
  Removed support for ActiveRecord and DataMapper and moved them into their own gems.
  Mack now works with an external orm gem.
  Working on breaking out orm support.
  Working on splitting out the ORM support into different gems.
  Added symbolize_keys to database config settings.
  ...

Conflicts:

	CHANGELOG

+++ lib/rendering/base.rb
@@ -41,7 +41,7 @@ module Mack
       private
       # Used to render a file from disk.
       def render_file(f, options = {})
-        options = {:is_partial => false, :ext => ".#{self.params(:format)}.erb", :dir => MACK_VIEWS}.merge(options)
+        options = {:is_partial => false, :ext => ".#{self.params(:format)}.erb", :dir => Mack::Configuration.views_directory}.merge(options)
         partial = f.to_s
         parts = partial.split("/")
         if parts.size == 1
@@ -53,6 +53,7 @@ module Mack
           parts[parts.size - 1] = "_" << parts.last if options[:is_partial]
           partial = File.join(options[:dir], parts.join("/") + options[:ext])
         end
+        raise Errno::ENOENT.new(partial) unless File.exists?(partial)
         return Mack::ViewBinder.render(File.open(partial).read, self.view_binder.controller, options)
       end
